requirements.txt
tweepy
feedparser
requests
python-dotenv
openai
